+++ CHANGELOG.md
@@ -7,3 +7,4 @@ HEAD
 * first commit
 * Add platformio library.json
 * Add first MQTTClient class and include ArduinoJson lib dep
+* Add release github action and build scripts